Counter application
Created a counter application in ReactJS that counts from 0 to 10. The application will have an "h1" tag which will display the counter and two buttons, one will increment the value and the other will decrement it. Also, the color of the "h1" counter will be changed based on the value of the counter. from 0 to 4 the color will be "green" from 5 to 9 the color will be "blue" for 10 the color will be "red". Deploy your result on Github.io and submit the link on Wise classroom.